datagen/requirements.txt
annotated-types==0.7.0
grpcio==1.65.1
grpcio-tools==1.65.1
pm4py==2.7.11.11
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
requests==2.32.3
